Calculate price

+++ Back-end/Oceanic/Oceanic.Common/Model/CalculatePriceViewModel.cs
@@ -6,7 +6,7 @@ namespace Oceanic.Common.Model
 {
     public class CalculatePriceViewModel
     {
-        public int weight { get; set; }
+        public decimal weight { get; set; }
         public int width { get; set; }
         public int height { get; set; }
         public int length { get; set; }
